add importing of scenarios

+++ src/scenario.h
@@ -318,7 +318,7 @@ typedef struct {
 	uint16 ride_ratings_num_brakes;
 	uint16 ride_ratings_num_reversers;
 	uint16 word_0138B5CE;
-	uint16 pad_0138B5D0[60];
+	uint8 pad_0138B5D0[60];
 	rct_ride_measurement ride_measurements[8];
 	uint32 next_guest_index;
 	uint16 grass_and_scenery_tilepos;
